refactor: shared ui components 타입 정의 및 분리
